# src/Juna.SKS.Package.Services/wwwroot/swagger-original.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Parcel Logistics Service",
    "contact": {
      "name": "SKS",
      "url": "http://www.technikum-wien.at/"
    },
    "version": "1.20.2"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/warehouse": {
      "get": {
        "tags": [ "warehouse-management" ],
        "summary": "Exports the hierarchy of Warehouse and Truck objects.\n",
        "operationId": "exportWarehouses",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/warehouse"
                }
              }
            }
          },
          "400": {
            "description": "An error occurred loading.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "No hierarchy loaded yet.",
            "content": {}
          }
        }
      },
      "post": {
        "tags": [ "warehouse-management" ],
        "summary": "Imports a hierarchy of Warehouse and Truck objects.\n",
        "operationId": "importWarehouses",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/warehouse"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully loaded.",
            "content": {}
          },
          "400": {
            "description": "The operation failed due to an error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      }
    },
    "/warehouse/{code}": {
      "get": {
        "tags": [ "warehouse-management" ],
        "summary": "Get a certain warehouse or truck by code",
        "operationId": "getWarehouse",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/hop"
                }
              }
            }
          },
          "400": {
            "description": "An error occurred loading.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "No hop with the specified id could be found."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "hop": {
        "required": [ "code", "description", "hopType", "locationCoordinates", "locationName", "processingDelayMins" ],
        "type": "object",
        "properties": {
          "hopType": {
            "type": "string"
          },
          "code": {
            "pattern": "^[A-Z]{4}\\d{1,4}$",
            "type": "string",
            "description": "Unique CODE of the hop."
          },
          "description": {
            "type": "string",
            "description": "Description of the hop."
          },
          "processingDelayMins": {
            "type": "integer",
            "description": "Delay processing takes on the hop."
          },
          "locationName": {
            "type": "string",
            "description": "Name of the location (village, city, ..) of the hop."
          },
          "locationCoordinates": {
            "$ref": "#/components/schemas/geoCoordinate"
          }
        },
        "example": {
          "code": "code",
          "locationName": "locationName",
          "processingDelayMins": 0,
          "hopType": "hopType",
          "description": "description",
          "locationCoordinates": {
            "lon": 1.4658129805029452,
            "lat": 6.027456183070403
          }
        },
        "discriminator": {
          "propertyName": "hopType"
        }
      },
      "warehouse": {
        "required": [ "code", "description", "hopType", "level", "locationCoordinates", "locationName", "nextHops", "processingDelayMins" ],
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/hop"
          },
          {
            "properties": {
              "level": {
                "type": "integer",
                "format": "int32"
              },
              "nextHops": {
                "type": "array",
                "description": "Next hops after this warehouse (warehouses or trucks).",
                "items": {
                  "$ref": "#/components/schemas/warehouse_nextHops"
                }
              }
            }
          }
        ]
      },
...
      "error": {
        "required": [ "errorMessage" ],
        "type": "object",
        "properties": {
          "errorMessage": {
            "type": "string",
            "description": "The error message."
          }
        }
      },
...
      "geoCoordinate": {
        "required": [ "lat", "lon" ],
        "type": "object",
        "properties": {
          "lat": {
            "type": "number",
            "description": "Latitude of the coordinate.",
            "format": "double"
          },
          "lon": {
            "type": "number",
            "description": "Longitude of the coordinate.",
            "format": "double"
          }
        },
        "example": {
          "lon": 1.4658129805029452,
          "lat": 6.027456183070403
        }
      },
...
      "warehouse_nextHops": {
        "required": [ "hop", "traveltimeMins" ],
        "type": "object",
        "properties": {
          "traveltimeMins": {
            "type": "integer"
          },
          "hop": {
            "$ref": "#/components/schemas/hop"
          }
        }
      }
    },
...
  }
}